% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_junc_count_info_recount2}
\alias{get_junc_count_info_recount2}
\title{Create the data frame of junctions from recount2 junction coverage files}
\usage{
get_junc_count_info_recount2(junction_coverage_file)
}
\arguments{
\item{junction_coverage_file}{Junction coverate file obtained from recount2.}
}
\value{
data frame for Junction_ID, Sample_ID and Junction_Count
}
\description{
Create the data frame of junctions from recount2 junction coverage files
}
