% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_prob_for_zibb}
\alias{get_prob_for_zibb}
\title{Calculate the density for the zero-inflated beta-binomial distribution}
\usage{
get_prob_for_zibb(params, s_n, s_k)
}
\arguments{
\item{params}{The vector of parameters for the zero-inflated beta-inomial distibution.}

\item{s_n}{The number of trials.}

\item{s_k}{The number of success.}
}
\value{
The density probability for the input arguments.
}
\description{
Calculate the density for the zero-inflated beta-binomial distribution
}
