% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_score}
\alias{get_score}
\title{Calculate the SF3B1ness scores for each sample}
\usage{
get_score(junc_counts_info)
}
\arguments{
\item{junc_counts_info}{data frame for Junction_ID, Sample_ID, Junction_Count.}
}
\value{
data frame for Sample_ID and SF3B1ness score.
}
\description{
Calculate the SF3B1ness scores for each sample
}
