% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample_ID_info}
\alias{Sample_ID_info}
\title{Information on recount2 samples}
\format{A data frame with 71,111 rows and 3 variables:
\describe{
  \item{X1}{Recount2 sample ID}
  \item{X2}{SRA project ID}
  \item{X3}{SRA run ID}
}}
\source{
\url{https://github.com/friend1ws/SF3B1ness/data-raw/sample_ids.tsv}
}
\usage{
Sample_ID_info
}
\description{
A dataset containing recount2 sample ID, SRA project ID and SRA run ID
for each sample analyzed in recount2.
}
\keyword{datasets}
