% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SF3B1_info}
\alias{SF3B1_info}
\title{Information on SF3B1 associated splicing junctions}
\format{A data frame with 575 rows and 12 variables:
\describe{
  \item{Junction_ID_Alt}{Recount2 junction ID for SF3B1 associated splicing junction}
  \item{Junction_ID_Ref}{Recount2 junction ID for corresponding normal splicing junction}
  \item{Junction_Key_Alt_GRCh38}{GRCh38 coordinates for SF3B1 associated splicing junction}
  \item{Junction_Key_Ref_GRCh38}{GRCh38 coordinates for corresponding normal splicing junction}
  \item{Junction_Key_Alt_GRCh37}{GRCh37 coordinates for SF3B1 associated splicing junction}
  \item{Junction_Key_Ref_GRCh37}{GRCh37 coordinates for corresponding normal splicing junction}
  \item{Alpha_0}{The 1st shape parameter of ZIBB distribution for wild type SF3B1}
  \item{Beta_0}{The 2nd shape parameter of ZIBB distribution for wild type SF3B1}
  \item{Pi_0}{The zero point-mass parameter of ZIBB distribution for wild type SF3B1}
  \item{Alpha_1}{The 1st shape parameter of ZIBB distribution for mutant type SF3B1}
  \item{Beta_1}{The 2nd shape parameter of ZIBB distribution for mutant type SF3B1}
  \item{Pi_1}{The zero point-mass parameter of ZIBB distribution for mutant type SF3B1}
}}
\source{
\url{https://github.com/friend1ws/SF3B1ness/data-raw/SF3B1_associated_junction_info.txt}
}
\usage{
SF3B1_info
}
\description{
A dataset containing various information (recount2 junction ID, coordinates
and parameter for zero-inflated beta-binomial (ZIBB) distributions) on
SF3B1 associated splicing junctions and their normal counter parts.
}
\keyword{datasets}
