% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sf3b1ness_recount2}
\alias{sf3b1ness_recount2}
\title{Obtain SF3B1ness score for recount2 junction coverate file}
\usage{
sf3b1ness_recount2(junction_coverage_file)
}
\arguments{
\item{junction_coverage_file}{Junction coverate file obtained from recount2.}
}
\value{
data frame for StudyID, Run ID and SF3B1ness score
}
\description{
Obtain SF3B1ness score for recount2 junction coverate file
}
