% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sf3b1ness_SJ}
\alias{sf3b1ness_SJ}
\title{Obtain SF3B1ness score for STAR SJ.out.tab files}
\usage{
sf3b1ness_SJ(SJ_file, ref = "hg19")
}
\arguments{
\item{SJ_file}{Junction file obtained through STAR alignment as SJ.out.tab file}
}
\value{
SF3B1ness score
}
\description{
Obtain SF3B1ness score for STAR SJ.out.tab files
}
